backend/requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
Flask-JWT-Extended==4.5.3
pymongo==4.5.0
Werkzeug==2.3.7
python-dotenv==1.0.0
Pillow==10.0.1